+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/ICustomFieldsProvider.kt
@@ -1,0 +1,15 @@
+// capture-sdk - bitdrift's client SDK
+// Copyright Bitdrift, Inc. All rights reserved.
+//
+// Use of this source code is governed by a source available license that can be found in the
+// LICENSE file or at:
+// https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
+
+package io.bitdrift.capture
+
+import io.bitdrift.capture.providers.Field
+
+/** Provides custom fields for emitted logs. */
+internal interface ICustomFieldsProvider {
+    fun customFields(): Array<Field>
+}